refactor(rust): open-by-default step typing and layered module structure
