--- rust/capture-pipelines-poc/src/demo.rs
@@ -1,183 +1,0 @@
-//! A small analytics pipeline wired from the framework pieces, reused by the
-//! `analytics_demo` integration test.
-//!
-//! `ParsedEvent` → [`Validate`] → [`ApplyQuota`]`.fail_open()` → [`ApplyRestrictions`]
-//! → (async `BatchAnnotate`, defined in the integration test) → verdicts. Every step
-//! declares only the capabilities it reads, and the composed effects struct
-//! [`AnalyticsFx`] is generated by [`compose_fx!`](crate::compose_fx).
-
-use crate::capability::{
-    HasDistinctId, HasEventName, HasLane, HasToken, Main, Restricted, Validated,
-};
-use crate::compose_fx;
-use crate::fx::{WarningEffects, WarningSink};
-use crate::outputs::AnalyticsOutputs;
-use crate::result::{NoOutputs, StepResult};
-use crate::step::{FallibleStep, Step};
-
-/// A minimally-parsed analytics event — only the fields the demo steps read.
-#[derive(Clone, Debug)]
-pub struct ParsedEvent {
-    /// Ingest token.
-    pub token: String,
-    /// Event name.
-    pub event: String,
-    /// Distinct id, if provided.
-    pub distinct_id: Option<String>,
-    /// Owning team (attribution for warnings).
-    pub team_id: u64,
-    /// Event timestamp (unix millis).
-    pub timestamp: i64,
-}
-
-impl HasToken for ParsedEvent {
-    fn token(&self) -> &str {
-        &self.token
-    }
-}
-impl HasEventName for ParsedEvent {
-    fn event_name(&self) -> &str {
-        &self.event
-    }
-}
-impl HasDistinctId for ParsedEvent {
-    fn distinct_id(&self) -> Option<&str> {
-        self.distinct_id.as_deref()
-    }
-}
-impl HasLane for ParsedEvent {
-    // Freshly-parsed events start on the main lane.
-    type Lane = Main;
-}
-
-// The demo effects struct: just the ingestion-warnings sink for now. Adding a second
-// concern later is one more field here — no framework change.
-compose_fx!(AnalyticsFx {
-    warnings: WarningSink,
-});
-
-/// Shape validation/normalization. Drops empty event names; warns (but continues) on
-/// suspiciously long tokens. Infallible — no I/O — so it needs no `fail_open()`.
-pub struct Validate;
-
-impl<Fx: WarningEffects> Step<ParsedEvent, Fx> for Validate {
-    type Out = Validated<ParsedEvent>;
-    type Outputs = NoOutputs;
-
-    fn apply(
-        &self,
-        event: ParsedEvent,
-        fx: &mut Fx,
-    ) -> StepResult<Validated<ParsedEvent>, NoOutputs> {
-        if event.event.is_empty() {
-            return StepResult::Drop {
-                reason: "invalid_event_name",
-            };
-        }
-        if event.token.len() > 40 {
-            fx.warn(
-                event.team_id,
-                "long_token",
-                format!("token length {}", event.token.len()),
-            );
-        }
-        StepResult::Continue(Validated::new(event))
-    }
-
-    fn name(&self) -> &'static str {
-        "validate"
-    }
-}
-
-/// An unexpected error from the quota limiter (e.g. Redis unavailable).
-#[derive(Debug)]
-pub struct QuotaError;
-
-/// Billing-quota check. Fallible: it returns `Err` when its backing store is
-/// unavailable, so it must be `fail_open()`-wrapped before joining the (infallible)
-/// capture chain. On the happy path it drops over-quota events.
-pub struct ApplyQuota {
-    /// A token whose lookup simulates an infra failure (returns `Err`).
-    pub failing_token: &'static str,
-}
-
-impl<In, Fx> FallibleStep<In, Fx> for ApplyQuota
-where
-    In: HasToken + HasEventName,
-{
-    type Out = In;
-    type Outputs = NoOutputs;
-    type Error = QuotaError;
-
-    fn apply(&self, event: In, _fx: &mut Fx) -> Result<StepResult<In, NoOutputs>, QuotaError> {
-        if event.token() == self.failing_token {
-            return Err(QuotaError); // limiter unavailable
-        }
-        if event.event_name() == "quota_blocked" {
-            return Ok(StepResult::Drop {
-                reason: "quota_limited",
-            });
-        }
-        Ok(StepResult::Continue(event))
-    }
-
-    fn name(&self) -> &'static str {
-        "apply_quota"
-    }
-}
-
-/// Event restrictions: redirects configured tokens to the DLQ or overflow output.
-/// This is the one demo step with a non-empty `Outputs` set, so it drives the
-/// [`OutputRegistry`](crate::outputs::OutputRegistry) completeness check.
-pub struct ApplyRestrictions {
-    /// Token forced to the DLQ.
-    pub dlq_token: &'static str,
-    /// Token forced to overflow.
-    pub overflow_token: &'static str,
-}
-
-impl<In, Fx> Step<In, Fx> for ApplyRestrictions
-where
-    In: HasToken + HasDistinctId + HasEventName,
-{
-    type Out = Restricted<In>;
-    type Outputs = AnalyticsOutputs;
-
-    fn apply(&self, event: In, _fx: &mut Fx) -> StepResult<Restricted<In>, AnalyticsOutputs> {
-        let target = {
-            let token = event.token();
-            if token == self.dlq_token {
-                Some(AnalyticsOutputs::Dlq)
-            } else if token == self.overflow_token {
-                Some(AnalyticsOutputs::Overflow)
-            } else {
-                None
-            }
-        };
-        match target {
-            Some(AnalyticsOutputs::Dlq) => StepResult::Redirect {
-                output: AnalyticsOutputs::Dlq,
-                preserve_key: false,
-            },
-            Some(AnalyticsOutputs::Overflow) => StepResult::Redirect {
-                output: AnalyticsOutputs::Overflow,
-                // hot key stays on its partition
-                preserve_key: true,
-            },
-            None => StepResult::Continue(Restricted::new(event, false, None)),
-        }
-    }
-
-    fn name(&self) -> &'static str {
-        "apply_restrictions"
-    }
-}
-
-/// Topic resolver for the demo pipeline's outputs — every variant is configured, so
-/// [`OutputRegistry::check`](crate::outputs::OutputRegistry::check) passes.
-pub fn analytics_topic_for(output: AnalyticsOutputs) -> Option<&'static str> {
-    match output {
-        AnalyticsOutputs::Overflow => Some("events_overflow"),
-        AnalyticsOutputs::Dlq => Some("events_dlq"),
-    }
-}
